move library files to under "/lib"
